Fix split package in aggs-matrix-stats module
The aggs-matrix-status module reuses a server package name for
utility methods. This commit moves these classes into the matrix
package name.
